Remove call for sysctl for processor name on Darwid
