+++ s_tui/helper_functions.py
@@ -19,9 +19,7 @@
 
 import os
 import logging
-import signal
 import platform
-import subprocess
 import re
 import csv
 import sys
@@ -29,34 +27,43 @@ import json
 import time
 
 from collections import OrderedDict
-from sys import exit
 
 __version__ = "1.0.0-beta"
 
+_DEFAULT = object()
+PY3 = sys.version_info[0] == 3
+POSIX = os.name == "posix"
+ENCODING = sys.getfilesystemencoding()
+if not PY3:
+    ENCODING_ERRS = "replace"
+else:
+    try:
+        ENCODING_ERRS = sys.getfilesystemencodeerrors()  # py 3.6
+    except AttributeError:
+        ENCODING_ERRS = "surrogateescape" if POSIX else "replace"
+
 
 def get_processor_name():
-    if platform.system() == "Windows":
-        return platform.processor()
-    elif platform.system() == "Darwin":
-        return subprocess.check_output(['/usr/sbin/sysctl', "-n",
-                                        "machdep.cpu.brand_string"]).strip()
-    elif platform.system() == "Linux":
-        command = "cat /proc/cpuinfo"
-        all_info = subprocess.check_output(command, shell=True).strip()
-        for line in all_info.split(b'\n'):
-            if b'model name' in line:
-                return re.sub(b'.*model name.*:', b'', line, 1)
-    return ""
+    """ Returns the processor name in the system """
+    if platform.system() == "Linux":
+        with open("/proc/cpuinfo", "rb") as cpuinfo:
+            all_info = cpuinfo.readlines()
+            for line in all_info:
+                if b'model name' in line:
+                    return re.sub(b'.*model name.*:', b'', line, 1)
+
+    return platform.processor()
 
 
-def kill_child_processes(parent_proc, sig=signal.SIGTERM):
+def kill_child_processes(parent_proc):
+    """ Kills a process and all its children """
     logging.debug("Killing stress process")
     try:
         for proc in parent_proc.children(recursive=True):
-            logging.debug('Killing' + str(proc))
+            logging.debug('Killing %s', proc)
             proc.kill()
         parent_proc.kill()
-    except(AttributeError):
+    except AttributeError:
         logging.debug('No such process')
         logging.debug('Could not kill process')
 
@@ -83,26 +90,26 @@ def output_to_csv(sources, csv_writeable_file):
 def output_to_terminal(sources):
     """Print statistics to the terminal"""
     results = OrderedDict()
-    for s in sources:
-        if s.get_is_available():
-            s.update()
-            results.update(s.get_summary())
+    for source in sources:
+        if source.get_is_available():
+            source.update()
+            results.update(source.get_summary())
     for key, value in results.items():
         sys.stdout.write(str(key) + ": " + str(value) + ", ")
     sys.stdout.write("\n")
-    exit()
+    sys.exit()
 
 
 def output_to_json(sources):
-    """Print statistics to the terminal"""
+    """Print statistics to the terminal in Json format"""
     results = OrderedDict()
-    for s in sources:
-        if s.get_is_available():
-            s.update()
-            source_name = s.get_source_name()
-            results[source_name] = s.get_sensors_summary()
+    for source in sources:
+        if source.get_is_available():
+            source.update()
+            source_name = source.get_source_name()
+            results[source_name] = source.get_sensors_summary()
     print(json.dumps(results, indent=4))
-    exit()
+    sys.exit()
 
 
 def get_user_config_dir():
@@ -110,7 +117,7 @@ def get_user_config_dir():
     Return the path to the user s-tui config directory
     """
     user_home = os.getenv('XDG_CONFIG_HOME')
-    if user_home is None or len(user_home) == 0:
+    if user_home is None or not user_home:
         config_path = os.path.expanduser(os.path.join('~', '.config', 's-tui'))
     else:
         config_path = os.path.join(user_home, 's-tui')
@@ -123,7 +130,7 @@ def get_user_config_file():
     Return the path to the user s-tui config directory
     """
     user_home = os.getenv('XDG_CONFIG_HOME')
-    if user_home is None or len(user_home) == 0:
+    if user_home is None or not user_home:
         config_path = os.path.expanduser(os.path.join('~', '.config',
                                                       's-tui', 's-tui.conf'))
     else:
@@ -156,29 +163,31 @@ def make_user_config_dir():
         try:
             os.mkdir(config_path)
             os.mkdir(os.path.join(config_path, 'hooks.d'))
-        except (OSError):
+        except OSError:
             return None
 
     return config_path
 
 
 def seconds_to_text(secs):
+    """ Converts seconds to a string of hours:minutes:seconds """
     hours = (secs)//3600
     minutes = (secs - hours*3600)//60
     seconds = secs - hours*3600 - minutes*60
     return "%02d:%02d:%02d" % (hours, minutes, seconds)
 
 
-def str_to_bool(s):
-    if s == 'True':
+def str_to_bool(string):
+    """ Converts a string to a boolean """
+    if string == 'True':
         return True
-    elif s == 'False':
+    if string == 'False':
         return False
-    else:
-        raise ValueError
+    raise ValueError
 
 
 def which(program):
+    """ Find the path of an executable """
     def is_exe(fpath):
         return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
 
@@ -193,6 +202,36 @@ def which(program):
                 return exe_file
 
     return None
+
+
+def _open_binary(fname, **kwargs):
+    return open(fname, "rb", **kwargs)
+
+
+def _open_text(fname, **kwargs):
+    """On Python 3 opens a file in text mode by using fs encoding and
+    a proper en/decoding errors handler.
+    On Python 2 this is just an alias for open(name, 'rt').
+    """
+    if PY3:
+        kwargs.setdefault('encoding', ENCODING)
+        kwargs.setdefault('errors', ENCODING_ERRS)
+    return open(fname, "rt", **kwargs)
+
+
+def cat(fname, fallback=_DEFAULT, binary=True):
+    """Return file content.
+    fallback: the value returned in case the file does not exist or
+              cannot be read
+    binary: whether to open the file in binary or text mode.
+    """
+    try:
+        with _open_binary(fname) if binary else _open_text(fname) as f_d:
+            return f_d.read().strip()
+    except (IOError, OSError):
+        if fallback is not _DEFAULT:
+            return fallback
+        raise
 
 
 DEFAULT_PALETTE = [
